modules/module-3-4/requirements.txt
flask
pandas
sqlalchemy
pymysql
mysql-connector-python
psycopg2-binary
python-dotenv
requests
streamlit

#  (tuỳ chọn - chỉ dùng khi cần)
openai
anthropic
google-generativeai